changed stats in models to be floats rather than integers

+++ xsoccer/playerstatistics/management/commands/build_playerstatistic_table_ALL_FILES.py
@@ -108,7 +108,7 @@ class Command(BaseCommand):
                             if stat_type == "formation_place":
                                 continue
 
-                            value = int(stats.text)
+                            value = float(stats.text)
 
                             playerstat = PlayerStatistic(
                                 game=Game.objects.get(uuid=game_uuid)
